+++ tests/generated-plugin/src/class-wpdtrt-test-widget.php
@@ -15,7 +15,7 @@
  * Since:
  *   0.7.16 - DTRT WordPress Plugin Boilerplate Generator
  */
-class WPDTRT_Test_Widget extends DoTheRightThing\WPDTRT_Plugin_Boilerplate\r_1_7_19\Widget {
+class WPDTRT_Test_Widget extends DoTheRightThing\WPDTRT_Plugin_Boilerplate\r_1_7_20\Widget {
 
 	/**
 	 * Constructor: __construct